Download synoptic data from PurpleAir

Download and parse synoptic data from the Purple Air network of particulate sensors.

The synoptic data provides a view of the entire Purple Air network and includes both metadata and recent PM2.5 averages for each deployed sensor.

pas_downloadParseRawData(baseUrl = "https://www.purpleair.com/json?all=true")

Arguments

baseUrl
base URL for synoptic data

Value

Dataframe of synoptic PurpleAir data.

References

json formatted PurpleAir data

See also

Examples

# \donttest{
library(AirSensor)

initializeMazamaSpatialUtils()

pas_raw <- pas_downloadParseRawData()

if ( interactive() ) {
  View(pas_raw[1:100,])
}
# }