Return IDs from filtered PurpleAir Synoptic objects

The incoming pas object is first filtered based on the values of stateCodes, patter, isOutside and isParent. The values associated with the "ID" column are then returned.

This function is useful for returning values associated with specific devices, which are represented by records with isParent = TRUE.

pas_getIDs(
  pas = NULL,
  pattern = ".*",
  idPattern = ".*",
  isOutside = TRUE,
  isParent = TRUE
)

Arguments

pas
PurpleAir Synoptic pas object.

pattern
Text pattern used to filter station labels.

idPattern
Text pattern used to filter deviceDeploymentID.

isOutside
Logical, is the sensor located outside?

isParent
Logical, is the record associated with a the A channel?

Value

Vector of values.

See also