Find Purple Air sensors within radial distance

Calculate Purple Air sensors within a specified distance from a target Purple Air sensor.

pas_within(pas, label = NULL, radius = "1 km")

Arguments

pas
a pa_synoptic dataframe

label
a target (centered) Purple Air sensor label

radius
a distance with unit (i.e "15 km")

Value

a pa_synoptic object

Details

radius should be a numeric string with provided metric unit, such as "250 m".

Examples

pas_within(example_pas,"SaratogaMiddleHighSchool", "1024 km")

#> # A tibble: 1,244 x 40
#>    ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>    <chr> <chr> <chr>            <chr>            <chr>           
#>  1 8304  MDAQ… outside          443822           R3OIPMBF7EIE8WBK
#>  2 8305  MDAQ… NA               443824           Q20INRIJMQG3W00Z
#>  3 10808 1-80… outside          496465           NJQEPVB0BHDCQKXA
#>  4 10809 1-80… NA               496467           U9BJQDBKB38741GM
#>  5 5460  1027… outside          385885           5CPPE62979927J8N
#>  6 5461  1027… NA               385887           8JBNGZA8AFWCTCCX
#>  7 2890  1092… outside          317401           R3CUAV2V4SX026VA
#>  8 2891  1092… NA               317403           N0C47KQZJV6AU8DI
#>  9 5416  121 … outside          385210           PAWKOGOH2HL1W1LB
#> 10 5417  121 … NA               385214           IR3SBZIJ3NHW1581
#> # … with 1,234 more rows, and 35 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, temperature <dbl>, humidity <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, pm25_current <dbl>, pm25_10min <dbl>,
#> #   pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>, pm25_1day <dbl>,
#> #   pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>