Aggregate PurpleAir Timeseries Object

Aggregate PurpleAir timeseries (pat) objects along its datetime axis. Temporal aggregation involves splitting a pat object into separate bins along its datetime axis. FUN is mapped to the pat numeric variables in each bin, which are then recombined into an aggregated pat object containing the same metadata as the incoming pat.

pat_aggregate(
  pat,
  FUN = function(x) {     mean(x, na.rm = TRUE) },
  unit = "minutes",
  count = 60
)

Arguments

pat
PurpleAir Timeseries pat object.

FUN
The function to be applied to each vector of numeric pat data.

unit
Character string specifying temporal units for binning.

count
Number of units per bin.

Value

Returns an aggregated pat object.

Details

FUN must operate on univariate numeric vectors and return a scalar value. Besides the data variable, no additional arguments will be provided to this function. This means that functions like mean and max will need to be wrapped in a function that specifies na.rm = TRUE. See the examples below.

Examples

library(AirSensor)

# Single day subset
pat <-
  example_pat %>%
  pat_filterDate(20180813, 20180814)

# Create aggregation functions
FUN_mean <- function(x) mean(x, na.rm = TRUE)
FUN_max <- function(x) max(x, na.rm = TRUE)
FUN_count <- function(x) length(na.omit(x))

# Hourly means
pat %>%
  pat_aggregate(FUN_mean) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#>               datetime    pm25_A    pm25_B pm1_0_atm_A pm1_0_atm_B pm25_atm_A
#> 1  2018-08-13 07:00:00  4.518222  5.016667  3.227778  3.713778   4.518222
#> 2  2018-08-13 08:00:00  4.989556  5.487778  3.504222  4.110000   4.989556
#> 3  2018-08-13 09:00:00  5.557556  5.868222  3.912222  4.442889   5.557556
#> 4  2018-08-13 10:00:00  6.377333  6.866889  4.490667  5.072667   6.377333
#> 5  2018-08-13 11:00:00  6.755111  7.245111  4.813556  5.426889   6.755111
#> 6  2018-08-13 12:00:00  6.762444  7.313778  4.712444  5.374667   6.762444
#> 7  2018-08-13 13:00:00  6.184667  7.026136  4.353111  5.127500   6.184667
#> 8  2018-08-13 14:00:00  7.115333  8.319111  5.042222  6.010667   7.115333
#> 9  2018-08-13 15:00:00  8.163556  9.521333  5.528000  6.829778   8.163556
#> 10 2018-08-13 16:00:00 13.711556 16.767778  8.020667 10.836000  13.711556
#> 11 2018-08-13 17:00:00 18.838444 22.071778 11.322000 14.155111  18.838889
#> 12 2018-08-13 18:00:00 19.213556 24.373333 11.580667 15.463556  19.226667
#> 13 2018-08-13 19:00:00 22.973333 28.197111 13.978667 18.001111  23.133111
#> 14 2018-08-13 20:00:00 27.460455 29.789556 16.430000 19.048000  27.987955
#> 15 2018-08-13 21:00:00 30.755778 31.096222 19.008444 20.271111  31.920000
#> 16 2018-08-13 22:00:00 30.974318 35.637111 19.080682 24.329111  32.546818
#> 17 2018-08-13 23:00:00 34.789778 39.832000 22.658444 29.032000  38.928889
#> 18 2018-08-14 00:00:00 38.628222 43.292444 26.390444 33.250000  45.974444
#> 19 2018-08-14 01:00:00 38.854889 43.918444 26.655111 33.722667  46.331111
#> 20 2018-08-14 02:00:00 39.989778 44.552889 27.655778 34.460667  48.225333
#> 21 2018-08-14 03:00:00 43.808222 47.332444 32.136444 38.483333  56.391556
#> 22 2018-08-14 04:00:00 49.525333 52.265556 39.775111 45.658889  70.538444
#> 23 2018-08-14 05:00:00 51.682889 54.769333 42.534444 48.586667  75.788222
#> 24 2018-08-14 06:00:00 53.074222 57.185556 44.307333 50.618000  78.899111
#>    pm25_atm_B pm10_0_atm_A pm10_0_atm_B
#> 1    5.016667   4.840667   5.565111
#> 2    5.487778   5.389333   5.962444
#> 3    5.868222   6.030000   6.422444
#> 4    6.866889   6.831111   7.559778
#> 5    7.245111   7.314222   7.889778
#> 6    7.313778   7.247333   8.255111
#> 7    7.026136   6.691333   7.767273
#> 8    8.319111   7.752667   9.082667
#> 9    9.521333   8.879556  10.556000
#> 10  16.767778  15.310667  19.040000
#> 11  22.093333  21.075556  24.896667
#> 12  24.424000  21.572667  27.618444
#> 13  28.734000  26.051556  32.933556
#> 14  31.290889  31.167500  35.882000
#> 15  33.115333  35.548222  37.968222
#> 16  40.224889  36.351591  46.316000
#> 17  47.996667  43.858667  55.044444
#> 18  55.170000  51.671111  63.042889
#> 19  56.634889  52.357778  65.100667
#> 20  57.924667  54.559111  66.460000
#> 21  64.838889  64.140000  74.554889
#> 22  77.528889  80.355111  89.233333
#> 23  82.791778  86.238000  95.168889
#> 24  86.806000  90.306000 100.264222

# Hourly maxes
pat %>%
  pat_aggregate(FUN_max) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#>               datetime pm25_A pm25_B pm1_0_atm_A pm1_0_atm_B pm25_atm_A pm25_atm_B
#> 1  2018-08-13 07:00:00   6.19   6.27      4.29      5.09       6.19       6.27
#> 2  2018-08-13 08:00:00   5.98   6.81      4.34      4.93       5.98       6.81
#> 3  2018-08-13 09:00:00   6.69   7.07      4.57      5.45       6.69       7.07
#> 4  2018-08-13 10:00:00   7.78   8.29      5.71      5.81       7.78       8.29
#> 5  2018-08-13 11:00:00   7.37   8.19      5.38      6.07       7.37       8.19
#> 6  2018-08-13 12:00:00   7.88   8.33      5.71      6.27       7.88       8.33
#> 7  2018-08-13 13:00:00   7.37   8.52      5.24      6.07       7.37       8.52
#> 8  2018-08-13 14:00:00   9.31  10.21      6.57      7.39       9.31      10.21
#> 9  2018-08-13 15:00:00  10.17  10.77      6.90      7.72      10.17      10.77
#> 10 2018-08-13 16:00:00  21.54  23.48     12.39     14.64      21.54      23.48
#> 11 2018-08-13 17:00:00  28.27  29.67     18.12     19.24      28.29      30.52
#> 12 2018-08-13 18:00:00  25.00  29.27     14.98     18.53      25.00      29.57
#> 13 2018-08-13 19:00:00  31.45  34.41     20.37     23.83      32.70      37.51
#> 14 2018-08-13 20:00:00  37.38  35.07     25.35     23.71      41.05      39.12
#> 15 2018-08-13 21:00:00  37.73  38.36     25.68     27.00      41.49      44.45
#> 16 2018-08-13 22:00:00  38.63  42.05     25.65     31.60      45.22      51.83
#> 17 2018-08-13 23:00:00  44.88  47.45     34.15     39.83      58.58      64.93
#> 18 2018-08-14 00:00:00  45.78  48.05     33.90     40.40      60.73      66.70
#> 19 2018-08-14 01:00:00  45.20  48.90     34.59     42.32      60.05      68.98
#> 20 2018-08-14 02:00:00  43.65  47.26     32.20     37.51      56.03      64.38
#> 21 2018-08-14 03:00:00  49.32  51.90     39.68     45.66      70.51      77.59
#> 22 2018-08-14 04:00:00  51.95  55.08     43.03     49.07      77.38      83.59
#> 23 2018-08-14 05:00:00  55.83  58.31     47.00     51.65      84.80      88.64
#> 24 2018-08-14 06:00:00  56.21  60.43     48.31     54.48      85.28      91.88
#>    pm10_0_atm_A pm10_0_atm_B
#> 1        7.14       6.76
#> 2        7.10       7.31
#> 3        7.76       7.48
#> 4        8.49       9.45
#> 5        8.13       9.17
#> 6        8.81       9.64
#> 7        7.73       9.59
#> 8       10.21      11.47
#> 9       11.10      12.62
#> 10      24.40      28.14
#> 11      32.66      34.05
#> 12      27.40      33.18
#> 13      35.72      42.02
#> 14      43.60      44.29
#> 15      45.29      51.05
#> 16      52.20      58.33
#> 17      65.10      73.76
#> 18      65.66      74.93
#> 19      67.78      80.02
#> 20      63.42      75.07
#> 21      80.02      87.85
#> 22      87.72      95.93
#> 23      94.13     103.05
#> 24      98.00     105.80

# Hourly counts
pat %>%
  pat_aggregate(FUN_count) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#>               datetime pm25_A pm25_B pm1_0_atm_A pm1_0_atm_B pm25_atm_A pm25_atm_B
#> 1  2018-08-13 07:00:00     45     45        45        45         45         45
#> 2  2018-08-13 08:00:00     45     45        45        45         45         45
#> 3  2018-08-13 09:00:00     45     45        45        45         45         45
#> 4  2018-08-13 10:00:00     45     45        45        45         45         45
#> 5  2018-08-13 11:00:00     45     45        45        45         45         45
#> 6  2018-08-13 12:00:00     45     45        45        45         45         45
#> 7  2018-08-13 13:00:00     45     44        45        44         45         44
#> 8  2018-08-13 14:00:00     45     45        45        45         45         45
#> 9  2018-08-13 15:00:00     45     45        45        45         45         45
#> 10 2018-08-13 16:00:00     45     45        45        45         45         45
#> 11 2018-08-13 17:00:00     45     45        45        45         45         45
#> 12 2018-08-13 18:00:00     45     45        45        45         45         45
#> 13 2018-08-13 19:00:00     45     45        45        45         45         45
#> 14 2018-08-13 20:00:00     44     45        44        45         44         45
#> 15 2018-08-13 21:00:00     45     45        45        45         45         45
#> 16 2018-08-13 22:00:00     44     45        44        45         44         45
#> 17 2018-08-13 23:00:00     45     45        45        45         45         45
#> 18 2018-08-14 00:00:00     45     45        45        45         45         45
#> 19 2018-08-14 01:00:00     45     45        45        45         45         45
#> 20 2018-08-14 02:00:00     45     45        45        45         45         45
#> 21 2018-08-14 03:00:00     45     45        45        45         45         45
#> 22 2018-08-14 04:00:00     45     45        45        45         45         45
#> 23 2018-08-14 05:00:00     45     45        45        45         45         45
#> 24 2018-08-14 06:00:00     45     45        45        45         45         45
#>    pm10_0_atm_A pm10_0_atm_B
#> 1          45         45
#> 2          45         45
#> 3          45         45
#> 4          45         45
#> 5          45         45
#> 6          45         45
#> 7          45         44
#> 8          45         45
#> 9          45         45
#> 10         45         45
#> 11         45         45
#> 12         45         45
#> 13         45         45
#> 14         44         45
#> 15         45         45
#> 16         44         45
#> 17         45         45
#> 18         45         45
#> 19         45         45
#> 20         45         45
#> 21         45         45
#> 22         45         45
#> 23         45         45
#> 24         45         45

# Alternative 10 minute aggregation (advanced users only - see details.)
pat %>%
  pat_aggregate(FUN_max, unit = "minutes", count = 10) %>%
  pat_extractData() %>%
  dplyr::select(1:9) %>%
  dplyr::slice(1:6)

#>              datetime pm25_A pm25_B pm1_0_atm_A pm1_0_atm_B pm25_atm_A pm25_atm_B
#> 1 2018-08-13 07:00:00   6.19   6.27      4.29      5.09       6.19       6.27
#> 2 2018-08-13 07:10:00   5.85   5.56      4.24      4.12       5.85       5.56
#> 3 2018-08-13 07:20:00   5.48   5.50      3.67      4.11       5.48       5.50
#> 4 2018-08-13 07:30:00   4.81   4.79      3.05      3.40       4.81       4.79
#> 5 2018-08-13 07:40:00   4.71   5.31      3.29      4.00       4.71       5.31
#> 6 2018-08-13 07:50:00   4.95   5.70      3.76      3.91       4.95       5.70
#>   pm10_0_atm_A pm10_0_atm_B
#> 1       7.14       6.76
#> 2       5.98       6.57
#> 3       5.95       6.55
#> 4       5.35       5.57
#> 5       4.88       6.02
#> 6       4.95       6.31