Daily state of health

This function combines the output of the State of Health (SoH) function arguments into a single tibble.

pat_dailySoH(pat = NULL, SoH_functions = c("PurpleAirSoH_dailyPctDC",
  "PurpleAirSoH_dailyPctReporting", "PurpleAirSoH_dailyPctValid",
  "PurpleAirSoH_dailyOtherFit", "PurpleAirSoH_dailyABFit"))

Arguments

pat
PurpleAir Timeseries pat object.

SoH_functions
Vector of function names. All the passed in functions must output tibbles with a daily `datetime` variable and must cover the same period of time.

Examples

SoH <-
  example_pat_failure_B %>%
  pat_dailySoH()
head(SoH)

#> # A tibble: 6 x 20
#>   datetime            pm25_A_pctDC pm25_B_pctDC pm25_A_pctRepor…
#>   <dttm>                     <dbl>        <dbl>            <dbl>
#> 1 2019-06-01 00:00:00            0            0             99.7
#> 2 2019-06-02 00:00:00            0            0             99.6
#> 3 2019-06-03 00:00:00            0            0             99.9
#> 4 2019-06-04 00:00:00            0            0             99.7
#> 5 2019-06-05 00:00:00            0            0             89.0
#> 6 2019-06-06 00:00:00            0            0             95.8
#> # … with 16 more variables: pm25_B_pctReporting <dbl>, pm25_A_pctValid <dbl>,
#> #   pm25_B_pctValid <dbl>, pm25_A_humidity_rsquared <dbl>,
#> #   pm25_A_temperature_rsquared <dbl>, pm25_B_humidity_rsquared <dbl>,
#> #   pm25_B_temperature_rsquared <dbl>, pm25_A_pm25_B_rsquared <dbl>,
#> #   pm25_A_pm25_B_slope <dbl>, pm25_A_pm25_B_intercept <dbl>,
#> #   temperature_pctDC <dbl>, temperature_pctReporting <dbl>,
#> #   temperature_pctValid <dbl>, humidity_pctDC <dbl>,
#> #   humidity_pctReporting <dbl>, humidity_pctValid <dbl>

#timeseriesTbl_multiplot(tbl, ylim = c(0,100))