Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <- 
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 44
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 39 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Ozone1 <chr>, pm25_current <dbl>,
#> #   pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>,
#> #   pm25_1day <dbl>, pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, deviceID <chr>, locationID <chr>, deviceDeploymentID <chr>,
#> #   airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2021-03-04 08:01:38 332430        19.4      33.5       40.5  35213   -76
#> 2 2021-03-04 08:03:38 332431        19.5      33         38.3  35215   -76
#> 3 2021-03-04 08:05:38 332432        18.3      31.4       35.3  35217   -76
#> 4 2021-03-04 08:07:38 332433        19.0      31.4       36.0  35219   -74
#> 5 2021-03-04 08:09:38 332434        19.6      31.2       34.4  35221   -74
#> 6 2021-03-04 08:11:38 332435        18.5      29.0       31.2  35223   -74
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2021-03-04 08:01:39 332445        3296.       931.       189.       23.5
#> 2 2021-03-04 08:03:39 332446        3264.       919.       174.       20.9
#> 3 2021-03-04 08:05:39 332447        3149.       872.       163.       21.8
#> 4 2021-03-04 08:07:39 332448        3076.       870        171.       20.3
#> 5 2021-03-04 08:09:39 332449        3266.       898.       152.       17.6
#> 6 2021-03-04 08:11:39 332450        3166.       880.       137.       12.7
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2021-03-04 08:01:41 332474        20.7      38.1       44.9  18408     0
#> 2 2021-03-04 08:03:41 332475        21.3      35.6       43.7  18240     0
#> 3 2021-03-04 08:05:41 332476        19.5      32.9       38.5  18240     0
#> 4 2021-03-04 08:07:41 332477        19.5      33.6       38.8  18240     0
#> 5 2021-03-04 08:09:41 332478        21.5      35.2       40.4  18408     0
#> 6 2021-03-04 08:11:41 332479        20.0      33.0       36.2  18240     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2021-03-04 08:01:42 332429        3301.       963.       187.       38.1
#> 2 2021-03-04 08:03:42 332430        3398.       981.       171.       31.3
#> 3 2021-03-04 08:05:42 332431        3092.       896.       160.       28.5
#> 4 2021-03-04 08:07:42 332432        3118.       912.       165.       27.2
#> 5 2021-03-04 08:09:42 332433        3373.       979.       165.       27.1
#> 6 2021-03-04 08:11:42 332434        3159.       921.       147.       24.0
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }