Load PurpleAir time series data for a week

A pre-generated PurpleAir Timeseries pat object will be loaded containing data for the most recent 7- or 45-day interval. Data are loaded from the archive set with either setArchiveBaseUrl() or setArchiveBaseDir() for locally archived files.

pat_loadLatest(id = NULL, label = NULL, pas = NULL, days = 7)

Arguments

id

PurpleAir sensor 'deviceDeploymentID'.

label

PurpleAir sensor 'label'.

pas

PurpleAir Synoptic pas object.

days

Number of days of data to include (7 or 45).

Value

A PurpleAir Timeseries pat object.

Note

Archive file names are generated with a unique "device-deployment" identifier by combining a unique location ID with a unique device ID. These deviceDeploymentID identifiers guarantee that movement of a sensor will result in the creation of a new time series.

Users may request a pat object in one of two ways:

1) Pass in id with a valid a deviceDeploymentID

2) Pass in both label and pas so that the deviceDeploymentID can be looked up.

See also

Examples

# \donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()
pat <- pat_loadLatest(label = "SCSB_07", pas = pas)
pat_multiPlot(pat)

}, silent = FALSE)
#> ERROR [2023-03-16 10:10:27] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230316.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230315.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230314.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230313.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230312.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230311.rda
#> ERROR [2023-03-16 10:10:28] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230310.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230309.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230308.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230307.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230306.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230305.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230304.rda
#> ERROR [2023-03-16 10:10:29] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230303.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230302.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230301.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230228.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230227.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230226.rda
#> ERROR [2023-03-16 10:10:30] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230225.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230224.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230223.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230222.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230221.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230220.rda
#> ERROR [2023-03-16 10:10:31] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230219.rda
#> ERROR [2023-03-16 10:10:32] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230218.rda
#> ERROR [2023-03-16 10:10:32] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230217.rda
#> ERROR [2023-03-16 10:10:32] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230216.rda
#> ERROR [2023-03-16 10:10:32] data file could not be loaded from: http://data.mazamascience.com/PurpleAir/v1/pas/2023/pas_20230215.rda
#> Error : Data file could not be loaded after 30 tries
# }