Display multiple plots on one page

A plotting function that uses ggplot2 to display multiple ggplot objects in a single pane. Can either be passed individual ggplot objects OR a pat object and a plot type. Typical usage would be to supply pat and use the plottype argument to quickly display preformatted plots.

"all": pm25_A, pm25_B, temperature, humidity

"pm25": PM2.5 from Channel A and Channel B

"pm25_over": PM2.5 from Channel A and Channel B in the same plot

"aux": auxillary data (temperature, humidity)

pat_multiplot(pat = NULL, plottype = "all", sampleSize = 5000,
  plotList = NULL, columns = NULL, a_size = 1, a_shape = 15,
  a_color = rgb(0.9, 0.25, 0.2), b_size = 1, b_shape = 15,
  b_color = rgb(0.2, 0.25, 0.9), t_size = 1, t_shape = 15,
  t_color = "black", h_size = 1, h_shape = 15, h_color = "black",
  alpha = 0.5)

Arguments

pat
Purple Air Timeseries "pat" object from createPATimeseriesObject()

plottype
Quick-reference plot types: "all", "aux", "pm25"

sampleSize
Either an integer or fraction to determine sample size

plotList
a list() of any number of ggplot objects to plot on a single pane

columns
Number of columns in the plot layout. Use NULL for defaults.

a_size
size of pm25_A points

a_shape
symbol to use for pm25_A points

a_color
color of pm25_A points

b_size
size of pm25_B points

b_shape
symbol to use for pm25_B points

b_color
color of pm25_B points

t_size
size of temperature points

t_shape
symbol to use for temperature points

t_color
color of temperature points

h_size
size of humidity points

h_shape
symbol to use for humidity points

h_color
color of humidity points

alpha
opacity of points

Value

Returns a ggplot object.

Note

Additional documentation of the multiplot algorithm is available at cookbook-r.com.