Aggregate PurpleAir Timeseries Data

Aggregate a data.frame along its datetime axis. Temporal aggregation involves splitting a data.frame object into separate bins along its datetime axis. FUN is mapped to the df numeric variables in each bin, which are then recombined into an aggregated data.frame.

patData_aggregate(
  df,
  FUN = function(df) {     mean(df$pm25_A + df$pm25_B, na.rm = TRUE) },
  unit = "minutes",
  count = 60
)

Arguments

df
Timeseries pat data, or timeseries data.frame with valid datetime column.

FUN
The function to be applied to each vector of numeric df.

unit
Character string specifying temporal units for binning.

count
Number of units per bin.

Value

Returns an aggregated data.frame object.

Details

This function intended for advanced users who wish to have more flexibility than the standard pat_aggregate() while aggregating timeseries data. FUN can operate and access all numeric vectors within the data.frame df.

Examples

library(AirSensor)

# Single day subset
pat <-
  example_pat %>%
  pat_filterDate(20180813, 20180814)

# Two Sample Student T-Test (advanced users only - see details.)
FUN_ttest <- function(data) {
  t.test(data$pm25_A, data$pm25_B)
}

t.testStats <-
  pat %>%
  pat_extractData() %>% # Note: Extract the timeseries data.frame
  patData_aggregate(FUN_ttest)

head(t.testStats)

#>              datetime statistic parameter      p.value                 conf.int
#> 1 2018-08-13 07:00:00 -3.870466  84.03118 0.0002139236   -0.7545392, -0.2423497
#> 2 2018-08-13 08:00:00 -4.785184  87.18799 6.910644e-06   -0.7051611, -0.2912834
#> 3 2018-08-13 09:00:00 -2.870907  87.55262  0.005131829 -0.52573065, -0.09560268
#> 4 2018-08-13 10:00:00 -3.427505   87.1562 0.0009318418   -0.7734414, -0.2056697
#> 5 2018-08-13 11:00:00 -4.980301  84.05045 3.343857e-06   -0.6856529, -0.2943471
#> 6 2018-08-13 12:00:00 -4.532544  87.43282  1.83922e-05   -0.7930868, -0.3095799
#>             estimate null.value alternative                  method
#> 1 4.518222, 5.016667          0   two.sided Welch Two Sample t-test
#> 2 4.989556, 5.487778          0   two.sided Welch Two Sample t-test
#> 3 5.557556, 5.868222          0   two.sided Welch Two Sample t-test
#> 4 6.377333, 6.866889          0   two.sided Welch Two Sample t-test
#> 5 6.755111, 7.245111          0   two.sided Welch Two Sample t-test
#> 6 6.762444, 7.313778          0   two.sided Welch Two Sample t-test
#>                     data.name
#> 1 data$pm25_A and data$pm25_B
#> 2 data$pm25_A and data$pm25_B
#> 3 data$pm25_A and data$pm25_B
#> 4 data$pm25_A and data$pm25_B
#> 5 data$pm25_A and data$pm25_B
#> 6 data$pm25_A and data$pm25_B