Time averages for PrupleAir time series

Function to flexibly aggregate or expand data frames by different time periods and calculating vector-averages for a PurpleAir time series object. This function should be useful in many circumstances where it is necessary to work with different time average data.

pat_aggregate(pat, period = "15 min", stats = "mean",
  parameter = NULL, dataThreshold = NULL, pprobs = NULL,
  quickStats = FALSE)

Arguments

pat
PurpleAir Timeseries "pat" object from pat_load()

period
The time period to average to. Can be "sec", "min", "hour", "day", "DSTday", "week", "month", "quarter" or "year". A number can also precede these options followed by a space (i.e. "2 day" or "37 min").

stats
The statistic to apply when aggregating the data; default is the mean. Can be one of "mean", "max", "min", "median", "frequency", "sd", "percentile", "tstats".

parameter
The variable of timeseries data to aggergate, such as "pm25_B", "temperature", etc.

dataThreshold
A that all data will be used in a particular period regardless of the number of values avaliable. Conversely, a value of 100 means that all data will need to be present to proceed, else it is recorded as NA.

pprobs
numeric vector of probabilities with values in [0,1]. Only valid when stats = "percentile"

quickStats
a logical that if TRUE will override stats parameter and return and a data frame of "mean", "sd", and "frequency".

Value

Returns a dataframe with a date in class POSIXct.

Examples

# NOT RUN {
# Hourly mean of Channel A PM2.5 density measurement
pat_aggregate(pat, "1 hour")
# }

# NOT RUN {
# Maximum weekly temperature
pat %>%
  pat_aggregate("temperature", "1 week", stats = "max")
# }