Aggregate PurpleAir Timeseries Object

Aggregate PurpleAir timeseries (pat) objects along its datetime axis. Temporal aggregation involves splitting a pat object into separate bins along its datetime axis. FUN is mapped to the pat numeric variables in each bin, which are then recombined into an aggregated pat object containing the same metadata as the incoming pat.

pat_aggregate(
  pat,
  FUN = function(x) {     mean(x, na.rm = TRUE) },
  unit = "minutes",
  count = 60
)

Arguments

pat
PurpleAir Timeseries pat object.

FUN
The function to be applied to each vector of numeric pat data.

unit
Character string specifying temporal units for binning.

count
Number of units per bin.

Value

Returns an aggregated pat object.

Details

FUN must operate on univariate numeric vectors and return a scalar value. Besides the data variable, no additional arguments will be provided to this function. This means that functions like mean and max will need to be wrapped in a function that specifies na.rm = TRUE. See the examples below.

Examples

library(AirSensor)

# Single day subset
pat <-
  example_pat %>%
  pat_filterDate(20180813, 20180814)

# Create aggregation functions
FUN_mean <- function(x) mean(x, na.rm = TRUE)
FUN_max <- function(x) max(x, na.rm = TRUE)
FUN_count <- function(x) length(na.omit(x))

# Hourly means
pat %>%
  pat_aggregate(FUN_mean) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#>               datetime pm25_A pm25_B temperature humidity pressure pm1_atm_A
#> 1  2018-08-13 07:00:00    4.5    5.0        64.7     61.4       NA       3.2
#> 2  2018-08-13 08:00:00    5.0    5.5        64.6     62.0       NA       3.5
#> 3  2018-08-13 09:00:00    5.6    5.9        64.3     62.7       NA       3.9
#> 4  2018-08-13 10:00:00    6.4    6.9        63.7     64.2       NA       4.5
#> 5  2018-08-13 11:00:00    6.8    7.2        64.0     62.9       NA       4.8
#> 6  2018-08-13 12:00:00    6.8    7.3        63.2     64.4       NA       4.7
#> 7  2018-08-13 13:00:00    6.2    7.0        62.3     66.3       NA       4.3
#> 8  2018-08-13 14:00:00    7.1    8.3        62.6     68.4       NA       5.0
#> 9  2018-08-13 15:00:00    8.2    9.5        64.2     68.1       NA       5.5
#> 10 2018-08-13 16:00:00   13.7   16.8        65.4     68.8       NA       8.0
#> 11 2018-08-13 17:00:00   18.8   22.1        67.5     65.2       NA      11.3
#> 12 2018-08-13 18:00:00   19.2   24.4        69.1     62.2       NA      11.6
#> 13 2018-08-13 19:00:00   23.0   28.2        70.6     59.6       NA      13.9
#> 14 2018-08-13 20:00:00   27.5   29.8        70.6     59.5       NA      16.3
#> 15 2018-08-13 21:00:00   30.8   31.1        70.4     60.8       NA      18.6
#> 16 2018-08-13 22:00:00   31.0   35.6        72.4     57.9       NA      18.7
#> 17 2018-08-13 23:00:00   34.8   39.8        74.2     54.9       NA      21.0
#> 18 2018-08-14 00:00:00   38.6   43.3        75.8     52.0       NA      23.2
#> 19 2018-08-14 01:00:00   38.9   43.9        74.8     53.4       NA      23.4
#> 20 2018-08-14 02:00:00   40.0   44.6        73.1     56.4       NA      24.1
#> 21 2018-08-14 03:00:00   43.8   47.3        71.9     59.7       NA      26.4
#> 22 2018-08-14 04:00:00   49.5   52.3        71.7     59.1       NA      29.9
#> 23 2018-08-14 05:00:00   51.7   54.8        71.0     58.7       NA      31.1
#> 24 2018-08-14 06:00:00   53.1   57.2        70.1     57.7       NA      31.9
#>    pm25_atm_A pm10_atm_A
#> 1         4.5        4.8
#> 2         5.0        5.4
#> 3         5.6        6.0
#> 4         6.4        6.8
#> 5         6.8        7.3
#> 6         6.8        7.2
#> 7         6.2        6.7
#> 8         7.1        7.8
#> 9         8.2        8.9
#> 10       13.7       15.3
#> 11       18.8       21.1
#> 12       19.2       21.6
#> 13       23.0       26.1
#> 14       27.5       31.3
#> 15       30.8       35.4
#> 16       31.0       36.2
#> 17       34.8       42.4
#> 18       38.6       48.1
#> 19       38.9       48.6
#> 20       40.0       50.3
#> 21       43.8       56.0
#> 22       49.5       64.2
#> 23       51.7       66.9
#> 24       53.1       68.7

# Hourly maxes
pat %>%
  pat_aggregate(FUN_max) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>               datetime pm25_A pm25_B temperature humidity pressure pm1_atm_A
#> 1  2018-08-13 07:00:00    6.2    6.3          66       62       NA       4.3
#> 2  2018-08-13 08:00:00    6.0    6.8          65       62       NA       4.3
#> 3  2018-08-13 09:00:00    6.7    7.1          65       64       NA       4.6
#> 4  2018-08-13 10:00:00    7.8    8.3          64       65       NA       5.7
#> 5  2018-08-13 11:00:00    7.4    8.2          64       64       NA       5.4
#> 6  2018-08-13 12:00:00    7.9    8.3          64       66       NA       5.7
#> 7  2018-08-13 13:00:00    7.4    8.5          63       68       NA       5.2
#> 8  2018-08-13 14:00:00    9.3   10.2          64       69       NA       6.6
#> 9  2018-08-13 15:00:00   10.2   10.8          65       70       NA       6.9
#> 10 2018-08-13 16:00:00   21.5   23.5          67       70       NA      12.4
#> 11 2018-08-13 17:00:00   28.3   29.7          68       67       NA      18.1
#> 12 2018-08-13 18:00:00   25.0   29.3          70       64       NA      15.0
#> 13 2018-08-13 19:00:00   31.4   34.4          71       61       NA      20.0
#> 14 2018-08-13 20:00:00   37.4   35.1          71       62       NA      23.4
#> 15 2018-08-13 21:00:00   37.7   38.4          72       62       NA      23.7
#> 16 2018-08-13 22:00:00   38.6   42.0          73       59       NA      23.0
#> 17 2018-08-13 23:00:00   44.9   47.5          75       56       NA      27.4
#> 18 2018-08-14 00:00:00   45.8   48.0          76       53       NA      27.3
#> 19 2018-08-14 01:00:00   45.2   48.9          75       55       NA      27.3
#> 20 2018-08-14 02:00:00   43.6   47.3          74       58       NA      26.6
#> 21 2018-08-14 03:00:00   49.3   51.9          72       60       NA      29.4
#> 22 2018-08-14 04:00:00   52.0   55.1          72       60       NA      31.3
#> 23 2018-08-14 05:00:00   55.8   58.3          71       59       NA      33.0
#> 24 2018-08-14 06:00:00   56.2   60.4          71       58       NA      33.5
#>    pm25_atm_A pm10_atm_A
#> 1         6.2        7.1
#> 2         6.0        7.1
#> 3         6.7        7.8
#> 4         7.8        8.5
#> 5         7.4        8.1
#> 6         7.9        8.8
#> 7         7.4        7.7
#> 8         9.3       10.2
#> 9        10.2       11.1
#> 10       21.5       24.4
#> 11       28.3       32.7
#> 12       25.0       27.4
#> 13       31.4       35.7
#> 14       37.4       43.5
#> 15       37.7       45.2
#> 16       38.6       49.1
#> 17       44.9       56.4
#> 18       45.8       57.0
#> 19       45.2       57.7
#> 20       43.6       55.6
#> 21       49.3       63.8
#> 22       52.0       67.3
#> 23       55.8       70.4
#> 24       56.2       72.0

# Hourly counts
pat %>%
  pat_aggregate(FUN_count) %>%
  pat_extractData() %>%
  dplyr::select(1:9)

#>               datetime pm25_A pm25_B temperature humidity pressure pm1_atm_A
#> 1  2018-08-13 07:00:00     45     45          45       45        0        45
#> 2  2018-08-13 08:00:00     45     45          45       45        0        45
#> 3  2018-08-13 09:00:00     45     45          45       45        0        45
#> 4  2018-08-13 10:00:00     45     45          45       45        0        45
#> 5  2018-08-13 11:00:00     45     45          45       45        0        45
#> 6  2018-08-13 12:00:00     45     45          45       45        0        45
#> 7  2018-08-13 13:00:00     45     44          45       45        0        44
#> 8  2018-08-13 14:00:00     45     45          45       45        0        45
#> 9  2018-08-13 15:00:00     45     45          45       45        0        45
#> 10 2018-08-13 16:00:00     45     45          45       45        0        45
#> 11 2018-08-13 17:00:00     45     45          45       45        0        45
#> 12 2018-08-13 18:00:00     45     45          45       45        0        45
#> 13 2018-08-13 19:00:00     45     45          45       45        0        45
#> 14 2018-08-13 20:00:00     44     45          44       44        0        45
#> 15 2018-08-13 21:00:00     45     45          45       45        0        44
#> 16 2018-08-13 22:00:00     44     45          44       44        0        45
#> 17 2018-08-13 23:00:00     45     45          45       45        0        45
#> 18 2018-08-14 00:00:00     45     45          45       45        0        44
#> 19 2018-08-14 01:00:00     45     45          45       45        0        45
#> 20 2018-08-14 02:00:00     45     45          45       45        0        45
#> 21 2018-08-14 03:00:00     45     45          45       45        0        45
#> 22 2018-08-14 04:00:00     45     45          45       45        0        45
#> 23 2018-08-14 05:00:00     45     45          45       45        0        45
#> 24 2018-08-14 06:00:00     45     45          45       45        0        45
#>    pm25_atm_A pm10_atm_A
#> 1          45         45
#> 2          45         45
#> 3          45         45
#> 4          45         45
#> 5          45         45
#> 6          45         45
#> 7          45         44
#> 8          45         45
#> 9          45         45
#> 10         45         45
#> 11         45         45
#> 12         45         45
#> 13         45         45
#> 14         44         45
#> 15         45         44
#> 16         44         45
#> 17         45         45
#> 18         45         44
#> 19         45         45
#> 20         45         45
#> 21         45         45
#> 22         45         45
#> 23         45         45
#> 24         45         45

# Alternative 10 minute aggregation (advanced users only - see details.)
pat %>%
  pat_aggregate(FUN_max, unit = "minutes", count = 10) %>%
  pat_extractData() %>%
  dplyr::select(1:9) %>%
  dplyr::slice(1:6)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>              datetime pm25_A pm25_B temperature humidity pressure pm1_atm_A
#> 1 2018-08-13 07:00:00    6.2    6.3          66       61       NA       4.3
#> 2 2018-08-13 07:10:00    5.8    5.6          65       61       NA       4.2
#> 3 2018-08-13 07:20:00    5.5    5.5          65       61       NA       3.7
#> 4 2018-08-13 07:30:00    4.8    4.8          65       62       NA       3.0
#> 5 2018-08-13 07:40:00    4.7    5.3          64       62       NA       3.3
#> 6 2018-08-13 07:50:00    5.0    5.7          64       62       NA       3.8
#>   pm25_atm_A pm10_atm_A
#> 1        6.2        7.1
#> 2        5.8        6.0
#> 3        5.5        6.0
#> 4        4.8        5.3
#> 5        4.7        4.9
#> 6        5.0        5.0