Aggregate PurpleAir Timeseries Object

Aggregate PurpleAir timeseries (pat) object along its datetime axis. Temporal aggregation involves splitting a pat object into separate bins along its datetime axis. FUN is mapped to the pat numeric variables in each bin, which are then recombined into an aggregated pat object containing the same metadata as the incoming pat.

pat_aggregate(
  pat,
  FUN = function(x) {     mean(x, na.rm = TRUE) },
  unit = "minutes",
  count = 60
)

Arguments

pat
PurpleAir Timeseries pat object.

FUN
The function to be applied to each vector of numeric pat data.

unit
Character string specifying temporal units for binning.

count
Number of units per bin.

Value

Returns an aggregated pat object.

Details

FUN must operate on univariate numeric vectors and return a scalar value. Besides the data variable, no additional arguments will be provided to this function. This means that functions like mean and max will need to be wrapped in a function that specifies na.rm = TRUE. See the examples below.

Examples

library(AirSensor)

# Single day subset
pat <- 
  example_pat %>% 
  pat_filterDate(20180813, 20180814)

# Create aggregation functions
FUN_mean <- function(x) mean(x, na.rm = TRUE)
FUN_max <- function(x) max(x, na.rm = TRUE)
FUN_count <- function(x) length(na.omit(x))

# Hourly means
pat %>%
  pat_aggregate(FUN_mean) %>% 
  pat_extractData() %>% 
  dplyr::select(1:9)

#> # A tibble: 24 x 9
#>    datetime            pm25_A pm25_B temperature humidity pressure pm1_atm_A
#>    <dttm>               <dbl>  <dbl>       <dbl>    <dbl>    <dbl>     <dbl>
#>  1 2018-08-13 07:00:00   4.52   5.02        64.7     61.4       NA      3.23
#>  2 2018-08-13 08:00:00   4.99   5.49        64.6     62         NA      3.50
#>  3 2018-08-13 09:00:00   5.56   5.87        64.3     62.7       NA      3.91
#>  4 2018-08-13 10:00:00   6.38   6.87        63.7     64.2       NA      4.49
#>  5 2018-08-13 11:00:00   6.76   7.25        64       62.9       NA      4.81
#>  6 2018-08-13 12:00:00   6.76   7.31        63.2     64.4       NA      4.71
#>  7 2018-08-13 13:00:00   6.18   7.03        62.3     66.3       NA      4.35
#>  8 2018-08-13 14:00:00   7.12   8.32        62.6     68.4       NA      5.04
#>  9 2018-08-13 15:00:00   8.16   9.52        64.2     68.1       NA      5.53
#> 10 2018-08-13 16:00:00  13.7   16.8         65.4     68.8       NA      8.02
#> # … with 14 more rows, and 2 more variables: pm25_atm_A <dbl>, pm10_atm_A <dbl>

# Hourly maxes
pat %>%
  pat_aggregate(FUN_max) %>% 
  pat_extractData() %>% 
  dplyr::select(1:9)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> # A tibble: 24 x 9
#>    datetime            pm25_A pm25_B temperature humidity pressure pm1_atm_A
#>    <dttm>               <dbl>  <dbl>       <dbl>    <dbl>    <dbl>     <dbl>
#>  1 2018-08-13 07:00:00   6.19   6.27          66       62       NA      4.29
#>  2 2018-08-13 08:00:00   5.98   6.81          65       62       NA      4.34
#>  3 2018-08-13 09:00:00   6.69   7.07          65       64       NA      4.57
#>  4 2018-08-13 10:00:00   7.78   8.29          64       65       NA      5.71
#>  5 2018-08-13 11:00:00   7.37   8.19          64       64       NA      5.38
#>  6 2018-08-13 12:00:00   7.88   8.33          64       66       NA      5.71
#>  7 2018-08-13 13:00:00   7.37   8.52          63       68       NA      5.24
#>  8 2018-08-13 14:00:00   9.31  10.2           64       69       NA      6.57
#>  9 2018-08-13 15:00:00  10.2   10.8           65       70       NA      6.9 
#> 10 2018-08-13 16:00:00  21.5   23.5           67       70       NA     12.4 
#> # … with 14 more rows, and 2 more variables: pm25_atm_A <dbl>, pm10_atm_A <dbl>

# Hourly counts
pat %>%
  pat_aggregate(FUN_count) %>% 
  pat_extractData() %>% 
  dplyr::select(1:9)

#> # A tibble: 24 x 9
#>    datetime            pm25_A pm25_B temperature humidity pressure pm1_atm_A
#>    <dttm>               <int>  <int>       <int>    <int>    <int>     <int>
#>  1 2018-08-13 07:00:00     45     45          45       45        0        45
#>  2 2018-08-13 08:00:00     45     45          45       45        0        45
#>  3 2018-08-13 09:00:00     45     45          45       45        0        45
#>  4 2018-08-13 10:00:00     45     45          45       45        0        45
#>  5 2018-08-13 11:00:00     45     45          45       45        0        45
#>  6 2018-08-13 12:00:00     45     45          45       45        0        45
#>  7 2018-08-13 13:00:00     45     44          45       45        0        44
#>  8 2018-08-13 14:00:00     45     45          45       45        0        45
#>  9 2018-08-13 15:00:00     45     45          45       45        0        45
#> 10 2018-08-13 16:00:00     45     45          45       45        0        45
#> # … with 14 more rows, and 2 more variables: pm25_atm_A <int>, pm10_atm_A <int>

# Alternative 10 minute aggregation (advanced users only - see details.)
pat %>%
  pat_aggregate(FUN_max, unit = "minutes", count = 10) %>%
  pat_extractData() %>%
  dplyr::select(1:9) %>%
  dplyr::slice(1:6)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> # A tibble: 6 x 9
#>   datetime            pm25_A pm25_B temperature humidity pressure pm1_atm_A
#>   <dttm>               <dbl>  <dbl>       <dbl>    <dbl>    <dbl>     <dbl>
#> 1 2018-08-13 07:00:00   6.19   6.27          66       61       NA      4.29
#> 2 2018-08-13 07:10:00   5.85   5.56          65       61       NA      4.24
#> 3 2018-08-13 07:20:00   5.48   5.5           65       61       NA      3.67
#> 4 2018-08-13 07:30:00   4.81   4.79          65       62       NA      3.05
#> 5 2018-08-13 07:40:00   4.71   5.31          64       62       NA      3.29
#> 6 2018-08-13 07:50:00   4.95   5.7           64       62       NA      3.76
#> # … with 2 more variables: pm25_atm_A <dbl>, pm10_atm_A <dbl>