Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <-
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 45
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 40 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Voc <dbl>, Ozone1 <chr>,
#> #   pm25_current <dbl>, pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>,
#> #   pm25_6hr <dbl>, pm25_1day <dbl>, pm25_1week <dbl>,
#> #   statsLastModifiedDate <dttm>, statsLastModifiedInterval <dbl>,
#> #   countryCode <chr>, stateCode <chr>, timezone <chr>, deviceID <int>,
#> #   locationID <chr>, deviceDeploymentID <chr>, airDistrict <chr>,
#> #   pwfsl_closestDistance <dbl>, pwfsl_closestMonitorID <chr>,
#> #   sensorManufacturer <chr>, targetPollutant <chr>, technologyType <chr>,
#> #   communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-07-23 07:01:35 178647        1.61      2.45       3.02     72   -95
#> 2 2020-07-23 07:03:28 178648        1.07      2.07       2.67     74   -96
#> 3 2020-07-23 07:05:42 178649        1.4       2.33       2.93     76   -95
#> 4 2020-07-23 07:07:35 178650        1.12      2.17       2.81     78   -96
#> 5 2020-07-23 07:09:30 178651        1.49      2.44       2.49     80   -95
#> 6 2020-07-23 07:11:31 178652        1.11      2.15       2.45     82   -96
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-07-23 07:01:37 178650         437.       125.       15.0       1.48
#> 2 2020-07-23 07:03:30 178651         387.       110.       17.1       1.45
#> 3 2020-07-23 07:05:45 178652         435.       124.       16.8       2.09
#> 4 2020-07-23 07:07:38 178653         389.       112.       14.2       1.98
#> 5 2020-07-23 07:09:34 178654         439.       123.       16.8       1.04
#> 6 2020-07-23 07:11:37 178655         378.       107.       14.4       2.42
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-07-23 07:01:40 178682        24.9      25.9       26.9  18592     0
#> 2 2020-07-23 07:03:32 178683        33.8      34.9       35.1  18088     0
#> 3 2020-07-23 07:05:46 178684        27.8      28.6       29.0  18592     0
#> 4 2020-07-23 07:07:40 178685        13.5      14.7       15.9  18088     0
#> 5 2020-07-23 07:09:36 178686        20.2      21.1       21.2  18088     0
#> 6 2020-07-23 07:11:39 178687        17.8      18.3       18.7  18256     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-07-23 07:01:42 178662        6501.      1537.       20.6       2.76
#> 2 2020-07-23 07:03:34 178663        9118.      2085.       14.0       1.27
#> 3 2020-07-23 07:05:48 178664        7300.      1690.       12.9       1.47
#> 4 2020-07-23 07:07:43 178665        3687.       864.       20.7       4.24
#> 5 2020-07-23 07:09:39 178666        5344.      1278.       16.3       1.93
#> 6 2020-07-23 07:11:40 178667        4733.      1131.       13.3       0.88
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }