Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <- 
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 44
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 39 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Ozone1 <chr>, pm25_current <dbl>,
#> #   pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>,
#> #   pm25_1day <dbl>, pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, deviceID <chr>, locationID <chr>, deviceDeploymentID <chr>,
#> #   airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-11-11 08:00:04 251181        14.4      22.2       23.7    753   -95
#> 2 2020-11-11 08:04:01 251182        16.1      26.4       27.9    757   -95
#> 3 2020-11-11 08:06:00 251183        16.1      24.5       26.2    759   -96
#> 4 2020-11-11 08:07:59 251184        17.1      26.3       28.7    761   -95
#> 5 2020-11-11 08:09:59 251185        17.1      27.9       30.5    763   -95
#> 6 2020-11-11 08:12:00 251186        18.1      29.4       32.1    765   -94
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-11-11 08:00:06 251203        2295.       640.       112.       7.32
#> 2 2020-11-11 08:02:53 251204        2511.       708.       136.      12.5 
#> 3 2020-11-11 08:04:04 251205        2550        714.       136.      10.1 
#> 4 2020-11-11 08:06:02 251206        2525.       708.       120.       8.32
#> 5 2020-11-11 08:08:01 251207        2606.       737.       133.       9.39
#> 6 2020-11-11 08:10:01 251208        2669.       753        145.      11.5 
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-11-11 08:00:07 251225        15.9      23.7       25.9  17952     0
#> 2 2020-11-11 08:02:57 251226        16.8      26.7       28.2  18928     0
#> 3 2020-11-11 08:04:06 251227        17.5      26         28.3  18256     0
#> 4 2020-11-11 08:06:07 251228        18.2      26.6       28.5  18424     0
#> 5 2020-11-11 08:08:03 251229        17.2      27.7       29.7  18256     0
#> 6 2020-11-11 08:10:02 251230        18.5      29.1       31    18256     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-11-11 08:00:09 251181        2496.       736.       124.       9.72
#> 2 2020-11-11 08:02:59 251182        2706.       794.       145.      10.9 
#> 3 2020-11-11 08:04:07 251183        2730.       799.       128.      10.4 
#> 4 2020-11-11 08:06:09 251184        2723.       814.       136.       9.73
#> 5 2020-11-11 08:08:05 251185        2780.       813.       147.      10.6 
#> 6 2020-11-11 08:10:04 251186        2904.       852.       146.      14.3 
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }