Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <-
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 44
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 39 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Ozone1 <chr>, pm25_current <dbl>,
#> #   pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>,
#> #   pm25_1day <dbl>, pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, deviceID <chr>, locationID <chr>, deviceDeploymentID <chr>,
#> #   airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-17 07:01:39 214065        50.0      119.       147.    693   -93
#> 2 2020-09-17 07:03:47 214066        49.4      120.       145.    695   -94
#> 3 2020-09-17 07:05:41 214067        50.1      120.       151.    697   -93
#> 4 2020-09-17 07:07:39 214068        48.6      117.       148.    699   -93
#> 5 2020-09-17 07:09:39 214069        48.5      118.       144.    701   -93
#> 6 2020-09-17 07:11:39 214070        48.8      118.       146.    703   -93
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-17 07:01:41 214072        9555.      2824.       981.       144.
#> 2 2020-09-17 07:03:50 214073        9432.      2805.       983.       143.
#> 3 2020-09-17 07:05:43 214074        9470.      2827.      1020.       152.
#> 4 2020-09-17 07:07:41 214075        9347.      2755        971.       149.
#> 5 2020-09-17 07:09:40 214076        9298.      2747.       964.       142.
#> 6 2020-09-17 07:11:41 214077        9288.      2762.       977.       144.
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-17 07:01:42 214095        54.2      122.       152.  17912     0
#> 2 2020-09-17 07:03:51 214096        54.9      119.       149.  18416     0
#> 3 2020-09-17 07:05:44 214097        55.0      120.       155.  17912     0
#> 4 2020-09-17 07:07:42 214098        53.3      122.       149.  17912     0
#> 5 2020-09-17 07:09:42 214099        53.8      120.       153.  18080     0
#> 6 2020-09-17 07:11:42 214100        53.5      118.       153.  18080     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-17 07:01:44 214066        9871.      2967.       918.       152.
#> 2 2020-09-17 07:03:53 214067        9787.      2965.       899.       145.
#> 3 2020-09-17 07:05:46 214068        9960.      2996.       914.       158.
#> 4 2020-09-17 07:07:44 214069        9753.      2921.       912.       159.
#> 5 2020-09-17 07:09:44 214070        9832.      2960.       928.       152.
#> 6 2020-09-17 07:11:44 214071        9638.      2914.       889.       158.
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }