Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <-
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 44
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 39 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Ozone1 <chr>, pm25_current <dbl>,
#> #   pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>,
#> #   pm25_1day <dbl>, pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, deviceID <chr>, locationID <chr>, deviceDeploymentID <chr>,
#> #   airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-08 07:01:03 207839        1.02      1.83       1.92   9584   -92
#> 2 2020-09-08 07:03:05 207840        0.88      1.07       1.35   9586   -91
#> 3 2020-09-08 07:05:03 207841        0.79      1.68       2.11   9588   -92
#> 4 2020-09-08 07:07:07 207842        1.12      1.92       2.22   9590   -91
#> 5 2020-09-08 07:09:05 207843        0.92      1.58       1.58   9592   -91
#> 6 2020-09-08 07:11:05 207844        1.04      1.65       1.65   9594   -90
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-08 07:01:05 207843         432.      115.       12.5        0.97
#> 2 2020-09-08 07:03:06 207844         372.       96.6       6.89       0.42
#> 3 2020-09-08 07:05:05 207845         366.      103.       14.4        1.4 
#> 4 2020-09-08 07:07:09 207846         397.      107.       16.2        1.08
#> 5 2020-09-08 07:09:06 207847         397       108.        8.32       0.83
#> 6 2020-09-08 07:11:06 207848         387.      100.        8.53       1.14
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-08 07:01:06 207870        2.09      2.67       2.67  18088     0
#> 2 2020-09-08 07:03:08 207871        3.31      3.55       3.55  18256     0
#> 3 2020-09-08 07:05:06 207872        2.55      3.02       3.7   18088     0
#> 4 2020-09-08 07:07:11 207873        2.86      3.25       3.54  18592     0
#> 5 2020-09-08 07:09:08 207874        2.93      3.52       3.67  18088     0
#> 6 2020-09-08 07:11:08 207875        2.97      3.41       3.41  18088     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-08 07:01:08 207842         574.       165.       8.61      0.46 
#> 2 2020-09-08 07:03:09 207843         792.       226.       7.28      0.55 
#> 3 2020-09-08 07:05:08 207844         679.       191.      13.4       1    
#> 4 2020-09-08 07:07:12 207845         720.       201.       9.71      0.93 
#> 5 2020-09-08 07:09:09 207846         734.       210.      13         0.570
#> 6 2020-09-08 07:11:09 207847         754.       215.       8.24      0.67 
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }