Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <-
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 45
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 40 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Voc <dbl>, Ozone1 <chr>,
#> #   pm25_current <dbl>, pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>,
#> #   pm25_6hr <dbl>, pm25_1day <dbl>, pm25_1week <dbl>,
#> #   statsLastModifiedDate <dttm>, statsLastModifiedInterval <dbl>,
#> #   countryCode <chr>, stateCode <chr>, timezone <chr>, deviceID <int>,
#> #   locationID <chr>, deviceDeploymentID <chr>, airDistrict <chr>,
#> #   pwfsl_closestDistance <dbl>, pwfsl_closestMonitorID <chr>,
#> #   sensorManufacturer <chr>, targetPollutant <chr>, technologyType <chr>,
#> #   communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-08-19 07:00:59 194111        0.24      0.76       0.93    191   -89
#> 2 2020-08-19 07:02:59 194112        0.38      1.09       1.47    193   -90
#> 3 2020-08-19 07:04:57 194113        0.79      1.66       1.82    195   -93
#> 4 2020-08-19 07:06:59 194114        0.52      1.09       1.25    197   -89
#> 5 2020-08-19 07:08:57 194115        0.63      1.35       1.9     199   -91
#> 6 2020-08-19 07:10:57 194116        0.98      1.47       1.73    201   -91
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-08-19 07:01:00 194124         238.       65.0       7.28       0.71
#> 2 2020-08-19 07:03:00 194125         262.       73.7       8.93       0.91
#> 3 2020-08-19 07:04:59 194126         332.       93.8      12.5        1.46
#> 4 2020-08-19 07:07:00 194127         299.       83.2       7.86       0.48
#> 5 2020-08-19 07:08:59 194128         289.       85.6      11.0        1.77
#> 6 2020-08-19 07:10:59 194129         320.       92.7      11.2        0.69
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-08-19 07:01:02 194145        0.33      0.89       1.14  18256     0
#> 2 2020-08-19 07:03:02 194146        0.61      1.02       1.02  18256     0
#> 3 2020-08-19 07:05:00 194147        0.59      0.83       0.83  18088     0
#> 4 2020-08-19 07:07:02 194148        0.78      1.16       1.55  18424     0
#> 5 2020-08-19 07:09:00 194149        0.73      1.18       1.54  18088     0
#> 6 2020-08-19 07:11:00 194150        1         1.27       1.27  18256     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-08-19 07:01:03 194119         225.       65.5       7.21       1.12
#> 2 2020-08-19 07:03:03 194120         258.       72.6       6.35       0.81
#> 3 2020-08-19 07:05:02 194121         280.       76.1       4.98       0.21
#> 4 2020-08-19 07:07:03 194122         301.       88.1       8.07       0.64
#> 5 2020-08-19 07:09:02 194123         298.       86.0      10.4        1.29
#> 6 2020-08-19 07:11:02 194124         305.       87.2       7.96       0.35
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }