Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <-
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 45
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 40 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Voc <dbl>, Ozone1 <chr>,
#> #   pm25_current <dbl>, pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>,
#> #   pm25_6hr <dbl>, pm25_1day <dbl>, pm25_1week <dbl>,
#> #   statsLastModifiedDate <dttm>, statsLastModifiedInterval <dbl>,
#> #   countryCode <chr>, stateCode <chr>, timezone <chr>, deviceID <chr>,
#> #   locationID <chr>, deviceDeploymentID <chr>, airDistrict <chr>,
#> #   pwfsl_closestDistance <dbl>, pwfsl_closestMonitorID <chr>,
#> #   sensorManufacturer <chr>, targetPollutant <chr>, technologyType <chr>,
#> #   communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-02 07:01:03 203527        0.97      1.33       1.43    944   -94
#> 2 2020-09-02 07:03:03 203528        0.98      1.79       1.81    946   -92
#> 3 2020-09-02 07:05:03 203529        1.04      1.64       2       948   -93
#> 4 2020-09-02 07:07:09 203530        1.14      1.91       2.22    950   -93
#> 5 2020-09-02 07:09:03 203531        1.32      2.02       2.02    952   -93
#> 6 2020-09-02 07:11:04 203532        0.66      1.36       1.66    954   -92
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-02 07:01:05 203534         348.       91.2       10.6       0.83
#> 2 2020-09-02 07:03:05 203535         343.       96.1       12.7       0.45
#> 3 2020-09-02 07:05:05 203536         376.       96.4       12.6       1.43
#> 4 2020-09-02 07:07:11 203537         372.      102.        15.9       1.31
#> 5 2020-09-02 07:09:04 203538         366.      104.        12.6       1.46
#> 6 2020-09-02 07:11:05 203539         319.       86.9       14.3       0.76
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2020-09-02 07:01:06 203559        1.33      2.3        2.37  18088     0
#> 2 2020-09-02 07:03:07 203560        1.35      2.12       2.12  18256     0
#> 3 2020-09-02 07:05:06 203561        1.02      2.16       2.24  18256     0
#> 4 2020-09-02 07:07:13 203562        1.44      2.02       2.18  18592     0
#> 5 2020-09-02 07:09:06 203563        1         1.98       2.02  18088     0
#> 6 2020-09-02 07:11:07 203564        1         1.76       1.76  18256     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2020-09-02 07:01:08 203531         389       113.        17.1       1   
#> 2 2020-09-02 07:03:08 203532         410.      113.        13.0       0.23
#> 3 2020-09-02 07:05:08 203533         363.      107.        18.0       1.29
#> 4 2020-09-02 07:07:14 203534         388.      117.        12.6       1.14
#> 5 2020-09-02 07:09:08 203535         333.       98.1       13.8       1.21
#> 6 2020-09-02 07:11:08 203536         307.       92.4       14.5       1.02
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }