Download PurpleAir timeseries data

Downloads timeseries data for a specific PurpleAir sensor from the ThingSpeak API and parses the content into individual dataframes. This function will always return dataframes with the appropriate columns even if no data are returned from ThingSpeak.

The returned list contains the following dataframes:

meta -- pas records for the specified sensor

A_PRIMARY -- channel A primary dataset

A_SECONDARY -- channel A secondary dataset

B_PRIMARY -- channel B primary dataset

B_SECONDARY -- channel B secondary dataset

These dataframes contain ALL data available from ThingSpeak for the specified sensor and time period.

See the references.

pat_downloadParseRawData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseUrl = "https://api.thingspeak.com/channels/"
)

Arguments

id
PurpleAir sensor 'deviceDeploymentID'.

label
PurpleAir sensor 'label'.

pas
PurpleAir Synoptic pas object.

startdate
Desired start time (ISO 8601).

enddate
Desired end time (ISO 8601).

timezone
Timezone used to interpret start and end dates.

baseUrl
Base URL for Thingspeak API.

Value

List containing multiple timeseries dataframes.

References

https://www2.purpleair.com/community/faq#!hc-sd-card-csv-file-header

Examples

# \donttest{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

pas <- pas_load()

pat_rawList <- 
  pat_downloadParseRawData(
    id = "78df3c292c8448f7_21257",
    pas = pas
  )

lapply(pat_rawList, head)

#> $meta
#> # A tibble: 2 x 44
#>   ID    label DEVICE_LOCATION… THINGSPEAK_PRIM… THINGSPEAK_PRIM…
#>   <chr> <chr> <chr>            <chr>            <chr>           
#> 1 21257 Linc… outside          648716           6D84VW7006XE07J9
#> 2 21258 Linc… NA               648718           NQWFIW5HLDO0FBFG
#> # … with 39 more variables: THINGSPEAK_SECONDARY_ID <chr>,
#> #   THINGSPEAK_SECONDARY_ID_READ_KEY <chr>, latitude <dbl>, longitude <dbl>,
#> #   pm25 <dbl>, lastSeenDate <dttm>, sensorType <chr>, flag_hidden <lgl>,
#> #   isOwner <int>, humidity <dbl>, temperature <dbl>, pressure <dbl>,
#> #   age <int>, parentID <chr>, flag_highValue <lgl>,
#> #   flag_attenuation_hardware <lgl>, Ozone1 <chr>, pm25_current <dbl>,
#> #   pm25_10min <dbl>, pm25_30min <dbl>, pm25_1hr <dbl>, pm25_6hr <dbl>,
#> #   pm25_1day <dbl>, pm25_1week <dbl>, statsLastModifiedDate <dttm>,
#> #   statsLastModifiedInterval <dbl>, countryCode <chr>, stateCode <chr>,
#> #   timezone <chr>, deviceID <chr>, locationID <chr>, deviceDeploymentID <chr>,
#> #   airDistrict <chr>, pwfsl_closestDistance <dbl>,
#> #   pwfsl_closestMonitorID <chr>, sensorManufacturer <chr>,
#> #   targetPollutant <chr>, technologyType <chr>, communityRegion <chr>
#> 
#> $A_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 uptime  rssi
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2021-03-25 07:00:57 347492        0        0.43        0.59   9974   -72
#> 2 2021-03-25 07:02:57 347493        0.28     0.79        0.91   9976   -74
#> 3 2021-03-25 07:04:58 347494        0.18     0.8         1.59   9978   -74
#> 4 2021-03-25 07:06:58 347495        0.34     1.25        1.93   9980   -72
#> 5 2021-03-25 07:08:58 347496        0        0.580       2      9982   -74
#> 6 2021-03-25 07:10:58 347497        0.07     0.95        1.4    9984   -74
#> # … with 3 more variables: temperature <dbl>, humidity <dbl>, pm2.5_atm <dbl>
#> 
#> $A_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2021-03-25 07:00:59 347509         190.       50.9       7.34       0.93
#> 2 2021-03-25 07:02:59 347510         228.       62.6       8.26       0.53
#> 3 2021-03-25 07:04:59 347511         259.       66.2       9.34       1.95
#> 4 2021-03-25 07:06:59 347512         247.       66.0      12.2        2.92
#> 5 2021-03-25 07:08:59 347513         184.       50.1      11.9        3.65
#> 6 2021-03-25 07:10:59 347514         222.       58.5      11.4        2.23
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 
#> $B_PRIMARY
#> # A tibble: 6 x 10
#>   created_at          entry_id pm1.0_cf1 pm2.5_cf1 pm10.0_cf1 memory  adc0
#>   <dttm>              <chr>        <dbl>     <dbl>      <dbl>  <dbl> <dbl>
#> 1 2021-03-25 07:01:00 347545       0.3        1.19       2.88  18240     0
#> 2 2021-03-25 07:03:00 347546       0.36       1.28       2.05  18240     0
#> 3 2021-03-25 07:05:00 347547       0.65       1.44       2.12  18240     0
#> 4 2021-03-25 07:07:01 347548       0.580      1.14       1.95  18408     0
#> 5 2021-03-25 07:09:01 347549       0.04       1.11       2.43  18240     0
#> 6 2021-03-25 07:11:01 347550       0.19       1.21       1.97  18240     0
#> # … with 3 more variables: pressure <dbl>, bsec_iaq <dbl>, pm2.5_atm <dbl>
#> 
#> $B_SECONDARY
#> # A tibble: 6 x 10
#>   created_at          entry_id counts_0.3 counts_0.5 counts_1.0 counts_2.5
#>   <dttm>              <chr>         <dbl>      <dbl>      <dbl>      <dbl>
#> 1 2021-03-25 07:01:01 347501         247.       74.0       13.8       3.77
#> 2 2021-03-25 07:03:01 347502         277.       79.1       14.0       2.16
#> 3 2021-03-25 07:05:02 347503         257.       77.0       13.7       2.56
#> 4 2021-03-25 07:07:02 347504         276.       79.4       12.0       1.53
#> 5 2021-03-25 07:09:02 347505         194.       57.9       15.7       3.32
#> 6 2021-03-25 07:11:02 347506         209.       60.2       11.9       2.67
#> # … with 4 more variables: counts_5.0 <dbl>, counts_10.0 <dbl>,
#> #   pm1.0_atm <dbl>, pm10.0_atm <dbl>
#> 

# }